backend/requirements.txt
# ===================
# FASTAPI & WEB
# ===================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# ===================
# BASE DE DONNÉES
# ===================
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ===================
# REDIS & CELERY
# ===================
redis==5.0.1
celery==5.3.6
flower==2.0.1

# ===================
# PLAYWRIGHT (Automatisation)
# ===================
playwright==1.41.0

# ===================
# STOCKAGE S3 (MinIO)
# ===================
boto3==1.34.25
minio==7.2.3

# ===================
# TELEGRAM
# ===================
python-telegram-bot==20.7

# ===================
# UTILITAIRES
# ===================
httpx==0.26.0
pillow==10.2.0
python-dotenv==1.0.0
tenacity==8.2.3
fake-useragent==1.4.0
